some clean up

+++ src/controller/ListItemController.go
@@ -15,8 +15,6 @@ type Controller struct {
 func (c *Controller) Index(w http.ResponseWriter, r *http.Request) {
 	items := c.Repository.GetItems()
 	data, _ := json.Marshal(items)
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	w.Header().Set("Access-Control-Allow-Origin", "*")
 	w.WriteHeader(http.StatusOK)
 	w.Write(data)
 	return
@@ -25,12 +23,8 @@ func (c *Controller) Index(w http.ResponseWriter, r *http.Request) {
 func (c *Controller) FindById(w http.ResponseWriter, r *http.Request)  {
 	params := mux.Vars(r)
 	id := params["id"]
-
 	item := c.Repository.GetItemById(id)
-
 	data, _ := json.Marshal(item)
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	w.Header().Set("Access-Control-Allow-Origin", "*")
 	w.WriteHeader(http.StatusOK)
 	w.Write(data)
 	return
@@ -40,7 +34,7 @@ func (c *Controller) FindById(w http.ResponseWriter, r *http.Request)  {
 func (c *Controller) Insert(w http.ResponseWriter, r *http.Request){
 	var i model.ListItem
 	if r.Body == nil{
-		http.Error(w, "Please send a request body", 400)
+		w.WriteHeader(http.StatusNoContent)
 		return
 	}
 	err := json.NewDecoder(r.Body).Decode(&i)
@@ -49,8 +43,8 @@ func (c *Controller) Insert(w http.ResponseWriter, r *http.Request){
 		return
 	}
 	c.Repository.InsertItem(i.Product,i.Quantity)
-	w.WriteHeader(http.StatusCreated)
 	res, _ := json.Marshal(i)
+	w.WriteHeader(http.StatusCreated)
 	w.Write(res)
 }
 
@@ -60,7 +54,7 @@ func (c *Controller) Update(w http.ResponseWriter, r *http.Request){
 	id := params["id"]
 
 	if r.Body == nil{
-		http.Error(w, "Please send a request body", 400)
+		w.WriteHeader(http.StatusNoContent)
 		return
 	}
 	err := json.NewDecoder(r.Body).Decode(&i)
@@ -68,14 +62,19 @@ func (c *Controller) Update(w http.ResponseWriter, r *http.Request){
 		http.Error(w, err.Error(), 400)
 		return
 	}
-
 	c.Repository.UpdateItem(id,i)
-
+	res, _ := json.Marshal(i)
+	w.WriteHeader(http.StatusAccepted)
+	w.Write(res)
 }
 
 func (c *Controller) Delete(w http.ResponseWriter, r *http.Request){
 	params := mux.Vars(r)
 	id := params["id"]
-
+	if id == ""{
+		w.WriteHeader(http.StatusNotAcceptable)
+		return
+	}
 	c.Repository.DeleteItem(id)
+	w.WriteHeader(http.StatusOK)
 }
